requirements.txt
flask
transformers
torch
mysql-connector-python
